requirements.txt
# Core dependencies
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.41
alembic==1.16.4
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.11.7
pydantic-settings==2.2.1

# AI and ML - Fixed versions for compatibility
openai==1.23.6
anthropic==0.8.0
langchain==0.1.0
langchain-openai==0.0.1
langchain-anthropic==0.0.1
langchain-community==0.0.9

# Vector and Memory
chromadb==0.4.24
redis==5.2.1
numpy==1.24.3
pandas==2.0.3

# Machine Learning
scikit-learn==1.3.0
scipy==1.11.1

# Phase 3 Advanced Intelligence
mlflow==2.8.1
plotly==5.17.0
dash==2.14.2
prophet==1.1.4

# Monitoring and Observability
prometheus-client==0.17.1
psutil==5.9.5

# AWS and Cloud
boto3==1.34.69
botocore==1.34.69

# Security
cryptography==41.0.7
fernet==1.0.1

# Utilities
requests==2.31.0
aiofiles==23.2.1
jinja2==3.1.2
markdown==3.5.1
ratelimit==2.2.1
bleach==6.1.0 